feat: add conflict file change type
