+++ helix-vcs/src/diff.rs
@@ -1,5 +1,4 @@
 use std::ops::Range;
-use std::path::{Path, PathBuf};
 use std::sync::Arc;
 
 use helix_core::Rope;
@@ -288,33 +287,6 @@ impl Diff<'_> {
                     None
                 }
             }
-        }
-    }
-}
-
-pub enum FileChange {
-    Untracked {
-        path: PathBuf,
-    },
-    Modified {
-        path: PathBuf,
-    },
-    Deleted {
-        path: PathBuf,
-    },
-    Renamed {
-        from_path: PathBuf,
-        to_path: PathBuf,
-    },
-}
-
-impl FileChange {
-    pub fn path(&self) -> &Path {
-        match self {
-            Self::Untracked { path } => path,
-            Self::Modified { path } => path,
-            Self::Deleted { path } => path,
-            Self::Renamed { to_path, .. } => to_path,
         }
     }
 }
